refactor(core): main.ts を gameplay/render に分割
Refs #8
